+++ src/App.tsx
@@ -5,6 +5,7 @@
 import { Routes, Route } from 'react-router-dom'
 import Layout from './components/Layout'
 import Home from './pages/Home'
+import Book from './pages/Book'
 import ShopByVehicle from './pages/ShopByVehicle'
 import ShopBySize from './pages/ShopBySize'
 import TireDetail from './pages/TireDetail'
@@ -15,6 +16,7 @@ function App() {
     <Routes>
       <Route element={<Layout />}>
         <Route path="/" element={<Home />} />
+        <Route path="/book" element={<Book />} />
         <Route path="/shop/vehicle" element={<ShopByVehicle />} />
         <Route path="/shop/size" element={<ShopBySize />} />
         <Route path="/tire/:id" element={<TireDetail />} />